part

+++ core/src/Domain/Contracts/Repositories/CarRepositoryInterface.php
@@ -6,7 +6,7 @@ use Rentacar\Domain\Entities\Car;
 
 interface CarRepositoryInterface
 {
-    public function getCarById(float $id): Car;
+    public function getCarById(int $id): Car;
     public function getAllCars(): array;
     public function createOrUpdateCar(Car $car): void;
     public function deleteCar(Car $car): void;
